[refactor/#48] Update import paths due to package restructuring
