+++ src/main/java/com/api/meetudy/chat/service/ChatService.java
@@ -13,9 +13,9 @@ import com.api.meetudy.global.response.exception.CustomException;
 import com.api.meetudy.global.response.status.ErrorStatus;
 import com.api.meetudy.global.utils.LeaderAccessValidator;
 import com.api.meetudy.member.entity.Member;
-import com.api.meetudy.study.group.entity.StudyGroup;
-import com.api.meetudy.study.group.entity.StudyGroupMember;
-import com.api.meetudy.study.group.repository.GroupRepository;
+import com.api.meetudy.group.entity.StudyGroup;
+import com.api.meetudy.group.entity.StudyGroupMember;
+import com.api.meetudy.group.repository.GroupRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
