+++ src/main/java/com/api/meetudy/global/init/DataInitializer.java
@@ -14,13 +14,13 @@ import com.api.meetudy.member.repository.MemberRepository;
 import com.api.meetudy.post.entity.Post;
 import com.api.meetudy.post.enums.PostCategory;
 import com.api.meetudy.post.repository.PostRepository;
-import com.api.meetudy.study.group.entity.StudyGroup;
-import com.api.meetudy.study.group.entity.StudyGroupMember;
-import com.api.meetudy.study.group.enums.GroupMemberStatus;
-import com.api.meetudy.study.group.enums.Location;
-import com.api.meetudy.study.group.enums.StudyCategory;
-import com.api.meetudy.study.group.repository.GroupMemberRepository;
-import com.api.meetudy.study.group.repository.GroupRepository;
+import com.api.meetudy.group.entity.StudyGroup;
+import com.api.meetudy.group.entity.StudyGroupMember;
+import com.api.meetudy.group.enums.GroupMemberStatus;
+import com.api.meetudy.group.enums.Location;
+import com.api.meetudy.group.enums.StudyCategory;
+import com.api.meetudy.group.repository.GroupMemberRepository;
+import com.api.meetudy.group.repository.GroupRepository;
 import com.api.meetudy.interest.entity.Interest;
 import com.api.meetudy.interest.entity.MemberInterest;
 import com.api.meetudy.interest.repository.InterestRepository;
