+++ src/main/java/com/api/meetudy/interest/repository/InterestRepository.java
@@ -1,6 +1,6 @@
 package com.api.meetudy.interest.repository;
 
-import com.api.meetudy.study.group.enums.StudyCategory;
+import com.api.meetudy.group.enums.StudyCategory;
 import com.api.meetudy.interest.entity.Interest;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
