+++ src/main/java/com/api/meetudy/member/dto/MemberUpdateDto.java
@@ -1,7 +1,7 @@
 package com.api.meetudy.member.dto;
 
-import com.api.meetudy.study.group.enums.Location;
-import com.api.meetudy.study.group.enums.StudyCategory;
+import com.api.meetudy.group.enums.Location;
+import com.api.meetudy.group.enums.StudyCategory;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.Pattern;
 import jakarta.validation.constraints.Size;
