+++ src/main/java/com/api/meetudy/mypage/service/MyPageService.java
@@ -8,12 +8,12 @@ import com.api.meetudy.member.dto.MemberDto;
 import com.api.meetudy.member.dto.MemberUpdateDto;
 import com.api.meetudy.member.entity.Member;
 import com.api.meetudy.member.mapper.MemberMapper;
-import com.api.meetudy.study.group.entity.StudyGroup;
-import com.api.meetudy.study.group.entity.StudyGroupMember;
-import com.api.meetudy.study.group.enums.GroupMemberStatus;
-import com.api.meetudy.study.group.enums.StudyCategory;
-import com.api.meetudy.study.group.repository.GroupMemberRepository;
-import com.api.meetudy.study.group.repository.GroupRepository;
+import com.api.meetudy.group.entity.StudyGroup;
+import com.api.meetudy.group.entity.StudyGroupMember;
+import com.api.meetudy.group.enums.GroupMemberStatus;
+import com.api.meetudy.group.enums.StudyCategory;
+import com.api.meetudy.group.repository.GroupMemberRepository;
+import com.api.meetudy.group.repository.GroupRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
